requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
dj-database-url==2.2.0
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==10.4.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
